Fix set of remarks
