--- release_info.json
@@ -1,1 +1,0 @@
-{"remote_url":"https://github.com/dataiku/dss-plugin-model-error-analysis","last_commit_id":"37f00f5502ed265badd9def60784d1ed981b39e2"}
