--- state.json
@@ -1,12 +1,0 @@
-{
-  "actions": [
-    {
-      "ts": 1676893273689,
-      "action": "INSTALL",
-      "user": "admin"
-    }
-  ],
-  "gitState": {
-    "enabled": false
-  }
-}
